Update `nodejs-npm-engine` to use new layer API
